Revert 156b6242966bb94b6664ed7d584e347f2bf3628b
We do not have to make an exception for localhost in the firewall, which
is the only reason which the original commit was made in the first
place: To "punch a hole" in the firewall for both localhost and a remote
proxy when a local SOCKS5-proxy was configured.

As it turns out, this never worked the way it was intended anyway due to
the semantics of `AllowEndpoint`.